fm_testing_tool/requirements.txt
jupyter

# from RTT (needs trimming / updates)
anytree
hypothesis
ipdb
ipython
ipywidgets
fileupload
matplotlib
pandas-datareader
parameterized
pathlib
pytest
qgrid
tabulate
